+++ full/chapters/10-interfaces/programs/Person.java
@@ -9,5 +9,5 @@ public class Person implements Ageable, Weighable {
     
     public int getAge() { return age; }
     
-    public int getWeight() { return weight; }   
+    public double getWeight() { return weight; }   
 }
