+++ full/chapters/11-inheritance/programs/Creature.java
@@ -8,5 +8,5 @@ public class Creature implements Ageable, Weighable {
     }
     
     public int getAge() { return age; }    
-    public int getWeight() { return weight; }
+    public double getWeight() { return weight; }
 }
